+++ source/extensions/omni.flux.property_widget_builder.model.usd/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [2.11.6] - 2024-06-26
+### Fixed
+- Texture preview windows sharing the same UI instance
+
 ## [2.11.5]
 ### Fixed
 - Fixed issue resetting color widget to default value